Add initial version of RingBuf struct
This change introduces the initial version of the crate. As it stands it
provides a RingBuf struct that implements a ring buffer with some
non-standard semantics. In particular, it assumes default constructible
elements and will always hold as many elements as the buffer can hold.
Newly added elements will silently evict older ones.