Add support for retrieving a RingBuf's front and back elements
Sometimes it is useful to directly retrieve special elements of a
collection, e.g., the first and last element. If nothing else, it
certainly would serve to mirror the functionality provided by VecDeque
and LinkedList, among others.
In any case, this change adds such support in the form of front and back
methods that retrieve the most recently added and least recently added
element, respectively. Along with that, and because we offer direct
access to the underlying slice, we also provide the means to retrieve
the front and back indexes, as they map into said slice.