Introduce RingBuf::from_vec constructor and ring_buf! macro
This change introduces a new constructor that allows users to create a
RingBuf object from an existing Vec. This constructor can be used to
effectively circumvent the requirement that T: Default, as is the case
for RingBuf::new.
Furthermore, using this new constructor, we add a new macro, ring_buf!,
that creates a RingBuf object using the same syntax as the standard vec!
macro.