Add support for indexing into RingBuf objects
Sometimes it is convenient to directly index into a RingBuf object in
to, say, change a value directly, without evicting an old one. This
change enables such use cases by adding support for regular bracket
style indexing syntax. The way it is envisioned, is that the index 0
always maps to the least recently added element (i.e., the back), and
index len-1 represents the most recently added element. Furthermore, we
automatically wrap around for accesses not within the bounds of the
buffer, alleviating the need for any out-of-bounds access handling.