Use unchecked but unsafe accesses when debug assertions are disabled
While right now there is no reason to suspect performance problems with
the crate because it is not really used in performance sensitive
contexts, there is also no reason to ignore supposedly low hanging fruit
from the start.
Repeated bounds checks in the iterator type we use as well as for other
accesses directly into the underlying slice are guaranteed to use valid
indexes -- we make sure of that. To that end, this change adjusts the
code to use unchecked (and unsafe) variants of accessor functions, at
least in builds that do not have debug assertions enabled. The thinking
is that this will strike a good balance between safety and performance.
Unfortunately, readability suffers a bit, but we pretty much consider
the crate to be code complete.