Use explicit checked methods for slice element access
We use the regular bracket indexing syntax for accessing elements of a
slice in various locations, at least when assertions are enabled. For
the sake of being explicit, but also to improve consistency with
respect to code using the unchecked access method variants, this change
switches over to using get method calls instead.