Introduce iterator! macro for defining RingIter
With upcoming changes we would like to also provide mutable access
through an iterator to RingBuf objects. For that to work we need to
define another type of iterator with only trivial differences over the
existing RingIter.
To prevent a bunch of duplicated code from showing up in the process,
this change introduces a macro for defining RingIter.