Use slice::as_ptr for accessing individual elements from iterator
We will soon introduce an additional iterator type that yields
mutable references to elements of the underlying ring buffer. Because
Rust only allows a single mutable reference to exist for any one object,
which includes the ring buffer slice, we cannot hand out multiple
mutable references to said slice. There are already some safe methods on
slices that could potentially enable our use case, but they would
require a major rework of our iterator.
The canonical solution to this problem is the usage of pointers. With
this change we embrace this solution and uses the slice::as_ptr method
as a base for retrieving references to individual elements.