Disallow creation of RingIterMut for containers with zero sized types
In our current implementation RingIterMut is a source of unsoundness,
because there is, after all, still a potential for handing out multiple
mutable references to the same object: namely when zero sized types come
into play. When a ring buffer of zero sized objects is being iterated
over, we will simply end up handing out references to the same object
over and over. In the case of mutable references (i.e., in a mutating
iterator context), this could mean more than one mutable reference to
the same object, if a client stores a yielded result.
With this change we disallow creation of a mutating iterator when the
ring buffer holds zero sized objects. There are ways to implement the
behavior correctly, but they would introduce additional complexity into
the code base. There is arguably little point in storing zero sized
objects in the ring buffer to begin with and so just disallowing the
usage of mutating iterators in such a scenario seems like a viable
option.