Add more test coverage for our iterator's size_hint
This change expands the test that we have in place for our iterator's
size_hint method. Specifically, we perform additional tests after
pushing various numbers of elements and also cover the usage of the
next_back method, for consuming elements from the iterator's end. While
at it, we also add an additional assertion to the respective code along
with an explanation as to why it should be checking a true invariant.