Run tests under Miri in CI
Miri is a tool that helps spot undefined behavior and other gotchas in
Rust code. It takes the form of an interpreter, which means that it can
only do so at runtime and when the code is actually being executed.
Given that we employ a few bits of unsafe code by now, it cannot hurt to
have Miri hooked up to confirm that what we are doing is sensible. To
that end, this change adds a stage that runs the tests using Miri to the
CI pipeline.