Rework front & back elements and pushing semantics
Redefine what our front and back elements represent to mirror
behavior/terminology of VecDeque more closely. So far our front was
pointing to the last element in a RingBuf created from a Vec. With this
change it refers to the first element. Furthermore, a push to the front
was writing to increasing addresses, which is counter to what pushing to
a VecDeque entails, which effectively prepends an element to the buffer.
We basically mirror this behavior but with wrapping semantics.