Adjust RingBuf::into_boxed_slice() to rearrange elements
So far the RingBuf::into_boxed_slice() method has been borderline
useless: the returned boxed slice had a layout that required knowledge
about where the front element was, depending on the use case.
With this change we make sure that the method rearranges elements in the
ring buffer so that the front forms the first element of the produced
slice and the back the last one.